kafka-components/kafka-producer-pos/requirements.txt
pandas
redis
kafka-python
pyarrow
